Moved NeRF test helpers into kornia.testing (part of tests reorg) + run pre-commit fixes
