+++ tests/nerf/test_nerf_solver.py
@@ -24,9 +24,12 @@ from kornia.nerf.nerf_model import NerfModelRenderer
 from kornia.nerf.nerf_solver import NerfSolver
 
 from testing.base import assert_close
-
-from tests.nerf.test_data_utils import create_random_images_for_cameras, create_red_images_for_cameras
-from tests.nerf.test_rays import create_four_cameras, create_one_camera
+from testing.nerf import (
+    create_four_cameras,
+    create_one_camera,
+    create_random_images_for_cameras,
+    create_red_images_for_cameras,
+)
 
 
 class TestNerfSolver:
